requirements.txt
# Flask and Web Framework Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7

# Open Policy Agent Integration
requests==2.31.0
flask-opa==1.0.0

# Browser Automation and Fingerprint Protection
selenium==4.15.2
fake-useragent==1.4.0
webdriver-manager==4.0.1

# Proxy and Network Handling
aiohttp==3.8.6
stem==1.8.2        # For Tor integration
PySocks==1.7.1     # SOCKS proxy support

# Security and Cryptography
cryptography==41.0.7
bcrypt==4.0.1

# Configuration and Environment
PyYAML==6.0.1
python-dotenv==1.0.0

# Testing Dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development Tools
black==23.9.1
flake8==6.1.0

PyJWT==2.8.0  # For JWT token support